fix sign in routes
